separates import and export logic
a 600 line file isn't helping things.  I'm trying to tease apart how this actually works and it makes it harder with everything all thrown in one place, especially with functions out of order and no clear dependency lines to see (i.e. module imports/exports).

There should be no logical (or code) changes whatsoever in this commit.
